Removed unnecessary scoreboard check
Used `at @s` instead of `at @a[scores={`(score)`)}]`, since it's smoother. This shouldn't break anything and may optimise things in the case of two players having the score on the same tick (as it'd run at both locations). This is also useful if you decide to slow the speed of checks.