Delete data/vanillog/advancements directory
there were a number of issues with using advancements, such as not being able to easily differentiate between a player opening a chest while crouched and placing a block off a chest while crouched. I'll just revert them back to objectives because that's easier and i feel it's simpler to just have everything be an objective.